requirements.txt
Pillow
django
djangorestframework
django-filter
drf-spectacular
drf-spectacular-sidecar
